Perform some refactoring on method and package names
