+++ AboutController.java
@@ -1,4 +1,4 @@
-import Utilities.VersionData;
+import util.VersionData;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
